Added custom notification for shotlock unlock
- Fixed slight focus bar issue having to do with precision.
- Added Focus orb
- Added configs for drops %
- Levelup commands will reset shotlocks as well

+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/ShotlockGUI.java
@@ -75,7 +75,7 @@ public class ShotlockGUI extends Screen {
 					{
 						matrixStack.translate((screenWidth - guiWidth * scale) - 19 * scale, (screenHeight - (guiHeight) * scale) - 8 * scale, 0);
 						matrixStack.scale(scale, scale, scale);
-						drawFocusCostBarTop(matrixStack, 0, 0, (int)(ClientEvents.focusGaugeTemp), playerData.getFocus(), scale);
+						drawFocusCostBarTop(matrixStack, 0, 0, (float)(ClientEvents.focusGaugeTemp), playerData.getFocus(), scale);
 					}
 					matrixStack.pop();
 					
@@ -83,7 +83,7 @@ public class ShotlockGUI extends Screen {
 					{
 						matrixStack.translate((screenWidth - guiWidth * scale) - 19 * scale, (screenHeight - (guiHeight) * scale) - 8 * scale, 0);
 						matrixStack.scale(scale, scale, scale);
-						drawFocusBarTop(matrixStack, 0, 0, (int)(ClientEvents.focusGaugeTemp), scale);
+						drawFocusBarTop(matrixStack, 0, 0, (float)(ClientEvents.focusGaugeTemp), scale);
 					}
 					matrixStack.pop();
 				}
